calculate, store, and display actual flow rate

+++ flow_server/db_writer_child.py
@@ -5,7 +5,7 @@ import sqlite3
 for line in sys.stdin:
     conn = sqlite3.connect('db.sqlite3')
     c = conn.cursor()
-    num = int(line)
+    num = float(line)
 
     with conn:
         SQL = """INSERT INTO display_flow_flow (val, timestamp)
